Unified ios and jvm message generation
