+++ grpc-mp-test/src/commonMain/proto/basicMessages.proto
@@ -110,7 +110,3 @@ message OneOfMessage {
     LongMessage field3 = 3;
   }
 }
-
-service TestService {
-  rpc testRpc (SimpleMessage) returns (SimpleRepeatedMessage);
-}
