+++ grpc-mp-test/src/jvmMain/kotlin/test.kt
@@ -1,6 +1,7 @@
 import com.google.protobuf.CodedOutputStream
 import com.google.protobuf.WireFormat
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.test.basic_messages.KMSimpleMessage
 
 fun main() {
-    CodedOutputStream.computeRawVarint32Size()
+    KMSimpleMessage("dralloooooooo")
 }
