+++ grpc-multiplatform-lib/src/iosJvmCommon/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/io/ParseException.kt
@@ -1,3 +1,3 @@
 package io.github.timortel.kotlin_multiplatform_grpc_lib.io
 
-expect class ParseException : Exception
+class ParseException : Exception()
