+++ grpc-multiplatform-lib/src/iosJvmCommon/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/MessageDeserializer.kt
@@ -1,0 +1,5 @@
+package io.github.timortel.kotlin_multiplatform_grpc_lib.message
+
+interface MessageDeserializer<T : KMMessage, K> {
+    fun deserialize(`data`: K): T
+}
