--- grpc-multiplatform-lib/src/iosMain/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/io/ParseException.kt
@@ -1,5 +1,0 @@
-package io.github.timortel.kotlin_multiplatform_grpc_lib.io
-
-actual class ParseException: Exception() {
-
-}
