--- grpc-multiplatform-lib/src/iosMain/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/MessageDeserializer.kt
@@ -1,7 +1,0 @@
-package io.github.timortel.kotlin_multiplatform_grpc_lib.message
-
-import platform.Foundation.NSData
-
-interface MessageDeserializer<T : KMMessage> {
-    fun deserialize(`data`: NSData): T
-}
