+++ plugin/src/main/java/io/github/timortel/kotlin_multiplatform_grpc_plugin/generate_mulitplatform_sources/generators/proto3_ios_file_writer.kt
@@ -1,7 +1,7 @@
 package io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators
 
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoFile
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.dsl.IOSDslBuilder
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.dsl.IosJvmDslBuilder
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.proto_file.IOSProtoFileWriter
 import java.io.File
 
@@ -17,5 +17,5 @@ fun writeIOSFiles(protoFile: ProtoFile, iosOutputDir: File) {
 //        .build()
 //        .writeTo(jvmOutputDir)
 
-    writeDSLBuilder(protoFile, IOSDslBuilder, iosOutputDir)
+    writeDSLBuilder(protoFile, IosJvmDslBuilder, iosOutputDir)
 }
